use audit_after rather than SCT.timestamp
It is most convenient to use the audit_after timestamp (despite it
having a small random delay in addition to an SCT's timestamp) because
it has already been vetted for future-dated timestamps.

The STH.timestamp field has not been vetted by the CTR, but anyone can
inspect these timestamps as they are available in the Tor consensus.