added Tor Browser details
Clarified that it is a three-hop circuit from the client's guard to the
sampled CTR, which means that the pool of such circuits must necessarily
be distinct (due to ending at middle relays rather than exit relays).

Also added a brief motivation as to why we should sample a new CTR in
case of failure (with forward ref to related security impact).