major refactor of red threat and extended auditor
- Switched order of extended log and extended auditor designs.  The idea
is that we first pitch how to "do more" the intuitive way without any
drastic ecosystem changes, which allows us to present the associated
costs: Tor consensus complexities, storage with threats of flooding,
etc.  Then, we can take a step back and show how things could be done
differently with "small but significant" changes to the CT landscape if
we accept that some, but not all, CT logs are honest in practise.
- Refactored extended auditor design according to the new structure and
base design.  In favour of readability, I added detailed steps in phase 2.
- Added placeholder subsections